# .codex-build/test-bin/Debug/net10.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/contracts": {
      "get": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "FirstName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "LastName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "BirthCountry",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProductId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "personId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsDescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Filters",
            "in": "query",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Contract"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}": {
      "get": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContractRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContractRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContractRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/beneficiaryClauseId": {
      "patch": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/locked": {
      "patch": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean"
              }
            },
            "text/json": {
              "schema": {
                "type": "boolean"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "boolean"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/available-supports": {
      "get": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "compartmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/recalculate-value": {
      "post": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/holdings": {
      "get": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/operations": {
      "get": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/contracts/{id}/rebuild-holdings": {
      "post": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/contracts/{id}/rebuild": {
      "post": {
        "tags": [
          "Contract"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/operations/contract/{contractId}": {
      "get": {
        "tags": [
          "Operation"
        ],
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OperationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contractId": {
            "type": "integer",
            "format": "int32"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "managementMode": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          },
          "currentValue": {
            "type": "number",
            "format": "double"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time"
          },
          "supports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FinancialSupportAllocationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ContractOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contractOptionTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customParameters": {
            "type": "string",
            "nullable": true
          },
          "optionType": {
            "$ref": "#/components/schemas/ContractOptionTypeDto"
          }
        },
        "additionalProperties": false
      },
...
      "CreateContractRequestDto": {
        "type": "object",
        "properties": {
          "contractNumber": {
            "type": "string",
            "nullable": true
          },
          "contractLabel": {
            "type": "string",
            "nullable": true
          },
          "contractType": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "locked": {
            "type": "boolean"
          },
          "dateSign": {
            "type": "string",
            "format": "date-time"
          },
          "dateEffect": {
            "type": "string",
            "format": "date-time"
          },
          "dateMaturity": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "postalAddress": {
            "type": "string",
            "nullable": true
          },
          "taxAddress": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "personId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "initialPremium": {
            "type": "number",
            "format": "double"
          },
          "totalPaidPremiums": {
            "type": "number",
            "format": "double"
          },
          "currentValue": {
            "type": "number",
            "format": "double"
          },
          "redemptionValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "paymentMode": {
            "type": "string",
            "nullable": true
          },
          "scheduledPayment": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "beneficiaryClauseId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "entryFeesRate": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "managementFeesRate": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "exitFeesRate": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "advisorComment": {
            "type": "string",
            "nullable": true
          },
          "hasAlert": {
            "type": "boolean"
          },
          "externalReference": {
            "type": "string",
            "nullable": true
          },
          "createdByUserId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "insuredPersonIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContractOptionDto"
            },
            "nullable": true
          },
          "compartments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompartmentDto"
            },
            "nullable": true
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OperationDetailsDto": {
        "type": "object",
        "additionalProperties": false
      },
      "OperationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contractId": {
            "type": "integer",
            "format": "int32"
          },
          "contractNumber": {
            "type": "string",
            "nullable": true
          },
          "contract": {
            "$ref": "#/components/schemas/ContractDto"
          },
          "type": {
            "$ref": "#/components/schemas/OperationType"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "operationDate": {
            "type": "string",
            "format": "date-time"
          },
          "executionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/OperationDetailsDto"
          },
          "advanceDetail": {
            "$ref": "#/components/schemas/AdvanceDetailDto"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OperationSupportAllocationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Entrer 'Bearer' suivi d'un espace et du token JWT",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}